% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpineR-package.R
\docType{package}
\name{alpineR-package}
\alias{alpineR}
\alias{alpineR-package}
\title{alpineR: Convert AlpineQuest trk files}
\description{
Convert AlpineQuest trk files to gpx and sf (simple feature) dataframes
}
\author{
\strong{Maintainer}: Dan Hicks \email{hicks.daniel.j@gmail.com}

}
\keyword{internal}
