% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{file_attributes}
\alias{file_attributes}
\alias{version}
\alias{set_ver}
\alias{pos}
\alias{set_pos}
\alias{reset_pos}
\alias{move_pos}
\title{File specification and read position attributes}
\usage{
version(x)

set_ver(object, version)

pos(x)

set_pos(object, position)

reset_pos(object)

move_pos(object, increment)
}
\arguments{
\item{object}{Object with attributes to be retrieved or set. Usually raw file contents, eg, output from \code{read_file_raw()}.}

\item{version}{Value to be set to the \code{"version"} attribute}

\item{position}{Value to be set for the \code{"position"} attribute}

\item{increment}{Value to be added to the \code{"position"} attribute}
}
\value{
Varies by function; either attribute value or \code{object} (invisibly)
}
\description{
Track file specification and read position
}
\details{
File specification version and read position are handled in an OOP style as object attributes.  These functions retrieve and set the values of these attributes.  Notably, they use \code{data.table::setattr()} to modify attributes from within functions by reference.
}
