% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{ggprofile}
\alias{ggprofile}
\title{Profile plot}
\usage{
ggprofile(ele, time, period = "second", every = 60)
}
\arguments{
\item{period}{Unit of time used for smooth}

\item{every}{How many time units to smooth together}
}
\value{
\code{ggplot2} plot
}
\description{
One-liner to create a profile plot (elevation over time) with smoothing
}
\details{
Elevations are converted to feet
}
