% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{gps_out}
\alias{gps_out}
\title{Export as GPX}
\usage{
gps_out(data_sf, path, groups = c(segment))
}
\arguments{
\item{data_sf}{Simple features dataframe to be exported}

\item{path}{Location to write GPX}

\item{group}{Locations will be ordered within groups defined by this variable(s)}
}
\description{
Write a parsed, tidied track to a GPX file
}
\details{
Assumes locations are ordered and column names are unchanged from the defaults. No data other than time and elevation are included.
}
