% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpineR-package.R
\name{imports}
\alias{imports}
\title{The Android app AlpineQuest uses a custom file specification to record GPS tracks.  This package provides functions for parsing these functions and exporting them as GPX files or sf (simple features) dataframes.}
\description{
The Android app AlpineQuest uses a custom file specification to record GPS tracks.  This package provides functions for parsing these functions and exporting them as GPX files or sf (simple features) dataframes.
}
