% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{locations}
\alias{locations}
\alias{get_locations}
\alias{get_location}
\alias{get_location_value}
\title{Locations}
\usage{
get_locations(source)

get_location(source)

get_location_value(source)
}
\arguments{
\item{source}{Raw track file contents to be parsed}
}
\value{
Nested list.  An individual location is a flat list of {key: value} pairs.
}
\description{
Parse locations
}
\details{
trk file version 4 enabled the data content of locations to be much more flexible, treating them as a container for one or more location values. Data will be read from the current read position of \code{source}, ie, \code{pos(source)}.
}
