% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segments}
\alias{segments}
\alias{get_segments}
\alias{get_segment}
\title{Segment structures}
\usage{
get_segments(source)

get_segment(source)
}
\arguments{
\item{source}{Raw track file contents to be parsed}
}
\value{
Nested list
}
\description{
Parse Segment structures
}
\details{
trk files can contain one or more segments.  Each segment can contain metadata and a series of locations.  Data will be read from the current read position of \code{source}, ie, \code{pos(source)}.
}
