% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{segments_to_sf}
\alias{segments_to_sf}
\title{Tidy a parsed trk file}
\usage{
segments_to_sf(parsed)
}
\arguments{
\item{parsed}{Parsed contents of a trk file, eg, as returned by \code{parse_trk()}}
}
\value{
See details
}
\description{
Converts the segments from a parsed trk file into a sf (simple features) dataframe.  One row = one location.  Column \code{segment} gives the segment index within the trk file.  File-level metadata, including the file basename, are attached as attributes.
}
