% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{to_linestring}
\alias{to_linestring}
\title{Aggregate tracks by date and segment}
\usage{
to_linestring(datasf, smoothing = 60)
}
\arguments{
\item{datasf}{sf dataframe to be aggregated}

\item{smoothing}{Amount of smoothing to be done in calculating total climb, in seconds}
}
\value{
sf dataframe with one row = one date-segment combination
}
\description{
Take the output of segments_to_sf() and aggregate to 1 row per segment + date
}
\details{
This generally assumes that the column names are unchanged from \code{segments_to_sf()}.  Geometry column is aggregated as type \code{LINESTRING}.  Columns are added for useful trk file metadata, including base filename and APQ version, along with segment-level statistics including total time and distance, elevation gain, and \href{https://www.nps.gov/shen/planyourvisit/how-to-determine-hiking-difficulty.htm}{Shenadoah Trail Difficulty}.
}
