% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{total_climb}
\alias{total_climb}
\title{Total elevation gain}
\usage{
total_climb(elevations, timestamps, .period = "second", .every = 60)
}
\arguments{
\item{elevations}{Vector of elevations}

\item{timestamps}{Vector of POSIX timestamps}

\item{.period}{Unit of time used for smooth}

\item{.every}{How many time units to smooth together}
}
\value{
Length-one numeric
}
\description{
Computes total elevation gained on a hike, with smoothing to reduce noise.
}
