% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{units}
\alias{units}
\alias{set_metres}
\alias{to_ft}
\alias{to_mi}
\title{Attach metres to a variable, and convert to feet or miles}
\usage{
set_metres(metres)

to_ft(metres)

to_mi(km)
}
\arguments{
\item{metres}{Numeric vector with values in metres}

\item{km}{Numeric vector with values in kilometers}
}
\description{
Attach metres to a variable, and convert to feet or miles
}
