% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waypoints.R
\name{waypoints}
\alias{waypoints}
\alias{get_waypoints}
\alias{get_waypoint}
\title{Waypoint structures}
\usage{
get_waypoints(source)

get_waypoint(source)
}
\arguments{
\item{source}{Raw track file contents to be parsed}
}
\value{
Nested list
}
\description{
Parse Waypoint structures
}
\details{
A waypoint can contain metadata and a single location. Data will be read from the current read position of \code{source}, ie, \code{pos(source)}.
}
